WIP: Add support for not-implmeented
Add support for not-implemented trap where a VOP can mark itself as
not implemented.  This causes a illtrap instruction to be inserted
followed by a branch always followed by a string (not necessarily nul
terminated) that represents the name of the VOP.

The signal handler currently catches the signal and sends prints out
the string and then continues.

Not yet debugged and definitely does not yet work.

We're just saving this in safe place for now.