Fix compiler warning about else
Add braces around the else clause that's indented as if it were part
of the else clause.  Inspection of the algorithm indicates that this
is probably the intended code and in this case doesn't change what
gets executed because the following statements would have been anyway.

Also verified a few random values that `(asin x)` and `(asin (float x
1w0))` produce the same values.  Only need to test 2^-27 <= x < 0.5.