Fix #41: Report proper process status
The main problem is that we weren't calling wait3 with WCONTINUED so
that we would be signaled when the process continues.  And we also
need to check that result of wait call was WCONTINUED>

Replace the wait3 routine with a C routine (prog_status) so we don't
have to deal with the OS-specific flags.  This function basically
returns what the lisp function wait3 did.

Use this function in GET-PROCESSES-STATUS-CHANGES.

 - runprog.c:
   - Add prog_status
 - run-program.lisp:
   - Use prog_status instead of wait3
 - issues.lisp:
   - Add basic test