Rename and add new reg assignments like RT
This is a first step at implementing call-named (as mentioned in issue
#76).  

As best as I can tell it would be really useful to have nfp and cname
as actual registers, which the RT port does.  Thus, remove nl1 and nl2
and add nfp and cname.  We didn't lose 2 of our non-descriptors
though. The RT port makes nfp and ocfp non-descriptors so we recover
that.  We'll see if this works out ok.

So, rename and add the registers, updating the list of non-descriptors
and descriptors.

Update arm-assem.s with these changes.  Note that we make nargs be r2,
the third arg to call_into_lisp. This slightly easier to read the
assembly code.

Update arm-lispregs.h to have the right register assignments and
names.

With this change running the kernel core gets as far as it used to
with the ame error message.