Bignum multiply without consing temp space
The current bignum multiplier creates temp space to hold the absolute
value of the bignums and then negates the result (in-place) at the
end.

Instead, use the algorithm from Hacker's Delight that pretends the
numbers are unsigned, does the unsigned multiply and finally corrects
the result.  No extra memory is needed for this.