Add function to get and set GC check options
Define GET-GC-ASSERTIONS to get the current GC assertions and
SET-GC-ASSERTIONS to set them.  These control the whether gc_assert is
called; whether a a pre-scan of generation 0 is done before GC;
whether we verify the heap after gc_free_heap is called from purify;
and the verify pointers to old-space when GCing generations greater
than or equal the given generation.

Update C code to change boolean variables to int's to make it simple
to to access from Lisp. (Don't have c-call:boolean.)