Clean up for fedora shell runner
We're not using the fedora docker anymore (because I can't get it to
work anymore and don't remember how I got it to work to begin with).
Thus, use the fedora shell executor to run CI so we don't need the
image or to wget lots of stuff because it's already installed.