Fix #80: Convert contribs to use ASDF to load
As it says, convert contribs to use ASDF to load them if possible.  Many
contribs already had an asd file so we basically just had to rename
foo.asd to contrib-foo.asd to keep backward compatibility with the old
contrib names. (And update the defsystem name to match.)

Added an asd file for packed-sse2.  Unix doesn't work, so it's left alone for now.