Refine features to include sse4 for popcnt
popcnt is part of sse4, not sse3/ssse3, so add feature detection for
sse4 to enable use of this instruciton.

sse4 and ppcnt was first available in the Nehalem architecture, Nov
2008 in the first gen of Core i7 and i5 processors.  The oldest
machines I have access to date from 2011.

Also remove :ssse3 because we don't currently define or use any ssse3
instructions.

Add a bit of documentation about popcnt.

Setup CI to build with a bootstrap file to get these instructions.