Revert changes to atanh
Still need the special case for real args to get the right branch
cuts.  Haven't yet figured out why the current code doesn't work when
the arg is a pure real.  Soemthing to do with computing y, but not
sure how to get that to work out without doing just explicitly
checking for real args.  In that case, it's just much easier to change
the arg to have the correct imaginary component to make it continuous
with the desired quadrant.