Allow using R12 in emit-ea
Currently R12 is not allowed as an index register (a carry over from
x86 because R12 is ESP when looking at the low 3 bits of the
register).  However, if we emit the appropriate REX prefix, then I
think R12 is ok to use here.  Just assume REX was emitted and allow
use of R12 for now.