Remove scavenging of read-only space.
`*SCAVENGE-READ-ONLY-SPACE*` is always set to `NIL`, and it's highly
unlikely any one ever sets it. I (rtoy) haven't done that in decades
at least.

So, remove this static symbol, and remove the C code that tests for
this to determine if the read-only space should be scavenged.

This requires a very simple cross-compile to remove the symbol.  See
boot-2020-04-1.lisp for very simple instructions.  We currently on do
this for x86 since we can't test on sparc right now.

Update CI to do a cross-compile.

Addresses #89