Round up length for boxed objects and simple vector
These objects should end on a double-word boundary, so round up the
len (from the header) to a double word before printing out the
contents.  If we don't round and we don't end on a boundary, we'll
probably just print an immediate word of 0.  Harmless, I think, but
we'll add a label that wastes space.