ldb prints out Unicode characters
When printing out a base-char, only the low 8 bits of the code were
used.  But with Unicode support, we need to take all the bits and
print them out.  For control codes we use the form "#\^x".  (Was
#\C-x, which isn't a valid supported character form.)  Ascii is
printed as normal "#\a", and everything else use uses "#\u+<hex>".

While we're at it, we also added special cases like #\Vt that are
listed in
https://cmucl.org/docs/cmu-user/html/Characters.html#Characters.

With this, we can print out all unicode characters in a form that can
be pasted back into lisp.