Fix handling of code header and symbol headers
Printing of the symbol header was broken because we printed out the
unbound marker as Lba, indicating an address.  But we just want the
raw value of the unbound marker

The code header consists of multiple function header objects so we
dump each of the objects now.

Finally a few changes on printing:
  * asm_word uses %08lx to the value
  * asm_lispobj prints out the address with %08lx to be consistent
    with asm_label.