asm_lispobj handles immediates
If the object is an even or odd immediate, just print out the value
since it's, duh, an immediate, not some kind of address or fixnum.

asm_symbol_header doesn't need the special case for the value slot
anymore.  (We can probably replace this with just asm_boxed as we did
before.)