Fix asm_closure_header and add table to print out header types
Instead of just dumping a closure header as a boxed structure, need to
be more careful to handle the closure function correctly.

Also added asmtab_types array to map the header type to the name of
the header.  This is mostly for making reading the assembly easier so
we have a name for each header type instead of memorizing the header
type value.