Dump the lisp assembly routines.
For now, we just write out .equ pseudo op for the address of each lisp
assembler routine so that the linker doesn't have any unresolved
values.

This is a workaround for now.  Ideally, these would be real
relocatable symbols at the right address in the read-only space so
that everything can be relocated.