Add prints and fix issue with breakpoint offset off by one.
When handling the breakpoint in `sigill_handler`, we were decrementing
the pc by one.  I guess that was for the old INT3 where the pc would
be at the next instruction.  With UD2, the pc is at the beginning of
the UD2 instruction.

With this change, setting breakpoints appear to work:
```
(break)
bp :start :function #'kernel:%srqt
0
(kernel:%sqrt 2d0)
*Breakpoint hit*
   [Condition of type SIMPLE-CONDITION]

Restarts:
  0: [CONTINUE] Return from BREAK.
  1: [ABORT   ] Return to Top-Level.

Debug  (type H for help)

(KERNEL:%SQRT 2.0d0)
```

Add some prints so we can see what's happening.

But removing breakpoints isn't working.