Use INT3 for breakpoints instead of UD2
Since lisp doesn't use INT3 anymore, we know that any INT3 instruction
is a breakpoint.  Hence, we also don't need to append the breakpoint
trap code.

Also need to update function_end_breakpoint_trap to use UD2 instead of
INT3.  This is needed because we need the trap code to tell lisp this
is a function end breakpoint.

Tested this by adding a start breakpoint for kernel:%sqrt.  Breakpoint
is hit and you can continue.   Also tested a function end breakpoint.
When the breakpoint is hit, we correctly see the return value(s), and
can continue fine.