Modify UD1 inst so it prints out registers correctly
Hack the definition of the ud1 instruction format so that the reg/mem
field is really a word register.  This then makes the pending
interrupt trap print out as "ud1 ecx, ecx" instead of "ud1 ecx, cl".

I don't know why the printer wanted to print cl, but `print-reg` was
passing `:byte` as the size (from the dstate `'width`).

This works around the issue.