Update the int instruction to disassemble int3
The int instruction would accept a code of 3 but would produce an int3
instruction.  However, the disassembly of this wasn't working since
the printer expected the opcode to be #b11001101 which didn't match
what was generated.  Hence, add a new printer to print int3 as "int
3".

Tested this by setting a breakpoint and disassembling the function to
see that the breakpoint is printed as "int 3" instead ".byte #xcc".

(Should we really make a separate int3 instruction?)