Define separate int and int3 instructions.
Note that `(inst int 3)` is two bytes long and is not converted to an
int3 instruction that is one byte long.

Tested by looking at the output of
`disassem::print-backend-inst-space` and also by inserting a
function-start breakpoint in `#'kernel:%sqrt` and disassembling the
function.  We see the `int3` instruction, and it is exactly one byte
long.