Add support for fallback resources in RUN-MOTIF-APPLICATION
At the public API level, it's a new keyword to RUN-MOTIF-APPLICATION.
At the protocol level, this is one extra object during the greeting.

Internally, this modifies message_read_string_table so that the
resulting StringTable has a NULL-terminated list of strings. This
seemed simpler than adding a whole new data transmission type for this
purpose.