Fix #121: fill-pointer-misc off by one for charpos
When computing the charpos of the stream, the location of the last
newline was subtracted from the fill-pointer of the output stream
string.  This points to where the next character would go, and
subtracting the location of the newline would be off by one.  We need
to decrement the result by 1 to get the true charpos.

Added a test for this.