64-bit support and clean up times
In unix.c define _LARGEFILE_SOURCE and _FILE_OFFSET_BITS so that we
can the 64-bit versions of stat and friends.  Also, for compatibility,
we don't return a struct timespec, but a long for the number of
seconds.  We ignore the nanosec part.

In unix.lisp, change the definition of dev-t on linux to be u-int64-t
since it seems we support the 64-bit integer type.