Don't use dev_t and friends in stat interface definition.
To simplify things, don't use dev_t and friends in the C interface.
Instead, use the largest integer type that any OS would want to use.
So dev_t becomes u_int64_t because Linux can return 64-bit integers
here.  This would work just fine for macos which uses 32-bit integers.

We are depending on the compiler to produce a warning or error if we
try to store a larger integer type into a smaller type.