Implement software-type in C
Add function software-type to misc.lisp, and initialize to
*software-type* to NIL so that (software-type) will set it
appropriately.

Add function os_software_type to os-common.c that returns the sysname
slot of struct utsname.   On Linux and macos, this value matches the
value that we previously returned.

In linux-os.lisp and bsd-os.lisp, comment out the code that sets
*software-type*.  (We need to do this for other OSes, still)