Move documentation function from misc.lisp to misc-doc.lisp
pclcom.lisp loads misc to get the documentation functions.  This
causes a problem when building cmucl with the new os_software_type
function because it's undefined on the first build because we haven't
defined it yet.

To simplify things, move the documentation functions from misc.lisp to
misc-doc.lisp so that pclcom can just load misc-doc to get just what
is needed.

This allows the build to proceed with os_software_type being undefined
in the first build.