Fix #136: ensure-directories-exist should return the given pathspec
The CLHS says

The primary value is the given pathspec so that this operation can be
straightforwardly composed with other file manipulation expressions.

The ansi tests interpret this to mean that the first arg should be
returned as is.

Previously, Cmucl was merging the pathspec with
`*default-pathname-defaults*` and returning that.