Fix #135: (unix-namestring ".") should be "."
The underlying issue is that `(pathname ".")` is `#P"./"`, which is true
on Unix since "." can only be a directory.  Changing this so that
`(pathname ".")` is `#P"."` involves making lots of changes, so for
now, just make `unix-namestring` prefix relative directories with "./"
instead of "".

This has a side-effect that (unix-namestring `(pathname
"abc/def.txt")` returns "./abc/def.txt" instead of "abc/def.txt".
Fortunately `(pathname "./abc/def.txt")` and `(pathname
"abc/def.txt")` return `equalp` pathname objects.