Use LANG envvar to set up external format and file name encoding
On startup use the LANG envvar to set up `*default-external-format*`
and set the system streams to use the specified external format.  The
filename encoding (`unix::*filename-encoding*`) is also set to the
same external format.

If the codeset specified by LANG is not known to cmucl, print a
warning, and leave `*default-external-format*` and friends unchanged.