Allow broadcast-stream and two-way-stream for stream-external-format
`stream-external-format` for a `broadcast-stream` is `:default`,
according to the CLHS.

For a `two-way-stream`, the CLHS doesn't say, so we'll return the
common external format if there is one, or `:default` if they're
different.

For a `synonym-stream`, the CLHS also doesn't say, so we'll just
return the format of the value of the synonym stream symbol.

Add tests for test different types of streams, except for
`broadcast-stream`.  The ansi-tests have tests for this.