Fix #155: Wrap help strings neatly
For the description for each switch neatly wrap the lines.

This isn't the best but works reasonably well.  From the description
string, extract all the words, where a word is some sequence of
characters separated by whitespace.

Then take that list and print each word followed by a space.  Insert a
newline as needed to wrap the line.