Add simple tests for external format of system streams
First, set LANG in `bin/run-tests.sh` so that we have a consistent
value, independent of the OSes setting.  Then we can reliably do tests
for external formats of the system terminal.

Second, verify that the external format for `sys:*stdin*`,
`sys:*stdout*`, `sys:*stderr*`, and `sys:*tty*` have the correct
external format.