Set *filename-encoding* to NIL in extfmts.lisp
Any file that needs to be opened by the extfmts code in extfmts.lisp
is done with `*filename-encoding*` set to `NIL`.  The names of the
files are always ascii so setting `*filename-encoding*` to `NIL`
bypasses any calls to `string-encode` to encode the filename.  This
simplifies processing of the files.

One additional minor tweak: in `%find-external-format`, also set
`*print-readably*` to `NIL` in `with-standard-io-syntax` because we
don't really need to be able to print everything readably in this
scope.