Fix up tests for filename encoding
In test issue.130, we set `*filename-encoding*` to `:utf-8`.  This
shouldn't be needed anymore, so remove it.

We also note here the in tests
issue.139-default-external-format-read-file and
issue.139-default-external-format-write-file, we were creating a
pathname consisting of Korean letters.  We didn't set the filename
encoding to utf-8 here when opening thes file so I don't know how that
ever worked.  But since the encoding is now :utf-8 it should be ok now.