Convert runtime strings to utf-16
The C runtime initializes several variables from either the
environment or the command line.  These are, of course, encoded in the
locale, so we need to convert these strings into utf-16 format.  Add
`decode-runtime-strings` to do just that.

This problem showed up when installing cmucl in /tmp/αβ and trying to
run lisp.  We get an error when printing the herald for the core file
that is being used because we call truename on the path.  Since the
path is encoded, we can't find the file.  We need to convert the path
to a utf-16 string that we can use.