Handle decode-runtime-strings and environment more carefully.
`decode-runtime-strings` needs to be more careful in converting the
strings from the C runtime.  For the command line parameters and the
environment list, we need to use the locale when converting to Lisp
strings.  But `*cmucl-lib*`, `*cmucl-core-path*`, and `*unidata-path*`
needs to use the file encoding to convert the C result to Lisp.

We also need to call `environment-init` after calling
`decode-runtime-strings` because the strings could have changed.  This
important for the "library:" search-list which contains file names.
Without this, the search-list is mangled so we can't find anything.

Finally, call `intl::setlocale` after the `environment-init` has been
called again because `intl::setlocale` needs the paths to the pot
files so the "library:" search-list has to be valid pathnames.
Previously, translations could be accessed early and error out because
the paths were not correct.

This was testing by installing in "/tmp/αβ" and running lisp.  Before
these changes, we errored out printing out the path to the core file
because the path to the core file was not properly decoded.  Now it
works.

Then we tried `(set-system-external-format :euc-kr)`.  We can find the
format implementation file correctly.

We also tested with
```
LANG=ko_KR.EUC_KR bin/lisp -noinit
```
This correctly loads up the euc-kr file and sets the external format.
There are no errors.  (But of course the printed path for the core
file is wrong because euc-kr can't handle the greek letters.