Fix bootstrap issues in pathname normalization for Darwin
We create a pathname in `filesys-init` during loading of the
kernel.core.  Unicode isn't ready at this time, so we can't do
pathname normalization yet.

Add `*enable-normalization*` to control this.  Defaults to `NIL` and
gets set to `T` in the initial-function.

This also requires that we load up the Unicode decomp and combining
tables before setting `*enable-normalization*` to `T`.

For testing, this is enabled on Linux where I do my development.
We'll have to reorder this when this is working.