Actually enable Darwin path normalization on Darwin.
Previously, we had this enabled on Linux for testing.  Now switch over
to Darwin.

Rename `*enable-normalization*` to
`*enable-darwin-path-normalization*` to make it clearer.  Add a
docstring too.

In save.lisp, only set it on darwin since it's not relevant to any
other system.

Fix up some compiler notes about unused variables in
`decompose-hangul-syllable`.