Disable issue.41.1 when running CI
This test was previously disabled only for Linux when running the CI.
However, it's now also failing when running the CI for Darwin.  Thus
disable it whenever we're running the CI.

I just manually tested this on my Linux and Mac boxes.  This test
passes without any problem.  Not sure what's going on.