Instead of calling %pathname-match-p call %%pathname-match-p
Here in `%enumerate-files`, we don't need the full capability of
`%pathname-match-p`.  The file we get is a string from reading from
the directory.  We can just use `parse-nametring` to get a pathname
object out of that.  Then it can be merged with the directory
pathname.  Thus, everything is a physical pathname suitable for
`%%pathname-match-p`.