Fix up test issues for pathnames.
With the updated pathname-match-p function, some of the existing
pathname tests were not correct.

For example "/tmp/foo.lisp" should NOT match "foo:*" because only the
name is `:WILD`.  The type is `NIL`, so "foo.lisp" doesn't match.  We
change the test to assert it fails and added a version where it should
pass.

There was also in trying to match different search-lists where we
forgot to call `pathname-match-p`.  Oops.