Fix #157: (directory "**/") only returns directories
Previously `(direrctory "**/")` returned all files in all
subdirectories.  The expectation was that only directories are
returned.  This behavior happens because `directory` would merge in
`:wild` for all the components of the pathname that aren't
filled (`NIL`).

Remove this magic merging the pathname.  Now, if you want files
returned, you need to specify `:wild` for those components.  So
`(directory "**/*.*")` returns all the files, just like how "**/" used
to.

Add a test to verify that only files are returned.