Use existing function to compute >= and <=
The code assumed if a0 >= b0 and a1 >= b1 implied a >= b for a
bigfloat.  But I forgot that 1.99999...w0 is represented as 2d0 and a
tiny negative number.  So `(<= 2w0 1.99999...w0)` would return false
because, while a0 <= b0, a1 <= b1 is false.

So for simplicity just use the obvious replacement that a <= b is the
same as a < b or a = b, which we already have.

[skip ci]