Add deftransform for >= and <= for integers.
See the comments in the code for this change.  I'm not sure this is
what we want, but the testsuite passes now when it was previously
failing.  The failure had something to do with utf-16-be external
format and there were warnings about `(>= foo 0)` having to use
`generic->=` because the first type was `unsigned-byte`.  This doesn't
happen anymore.