Fix #172: Declare pathname-match-p to match implementation
The function `pathname-match-p` returns either `NIL` or a `pathname`.
But it was declared to return a `boolean`.  That is `NIL` or `T`.  The

The CLHS says it returns a generalized boolean so returning a pathname
is allowed.  Let's just update the `defknoen` for `pathname-match-p`
to reflect what we actually do.