Fix #171: Readably print pathnames with :unspecific
If a pathname contains a component with `:unspecific`, don't use the
namestring to print the pathname.  Use cmucl's extension `#P(...)` to
print the pathname readably.

We also tighten the algorithm so that the extension is not used for
printing pathnames that have an :unspecific component with lisp
patterns or search-lists.  Patterns and search-lists don't have a
readable syntax.