Fix #192:  Print radix marker in disassemblies; adjust note column
Change the default for `:radix` in `disassem:disasemble` from
`*print-radix*` to `T` so that the radix marker is printed by
default.  `disassemble` calls `disassem:disassemble` with no keyword
args specified.

Also instructions on x86 can be long so that the default
`*note-column*` is too small to have all the notes line up neatly.
Increase it for x86.  For other architectures, the default is good
enough (I think) so no need to increase it.