Fix #228: Updatea ansi-tests from upstream
First, change `.gitlab-ci.yml` not to change to the
`rtoy-cmucl-expected-failures` branch so that CI will run and generate
the list of failures for us.

Second, create a new branch in cmucl/ansi-tests to add our list of
expected failures.

Third, update this yaml file to change to the correct branch and rerun
CI.  There should be no unexpected failures.