Add tests for different foreign integer return types
Add a C file, test-return.c, that has functions that return different
length integer types.

Compile it in run-tests.sh so that the tests can load it.

Add tests that we get the right values from the functions.  Only the
signed integer types are tested right now.  We need to add the
unsigned tests.