Handle the case of NIL in list2.
When list2 is in the hash table and it contains NIL and list1 also
contains NIL, we were erroneously including NIL in the difference.
This was because we weren't checking the second return value for
gethash.

Also, for consistency, don't reverse the set of differences.  This
makes output from the hash-table version match the list version.

Finally, add a defparameter to specify the length threshold for
converting a list to hashtable.  This is mostly for testing to
determine an appropriate value.