Remove arg from test function and set the value directly.
As suggested by @cshapiro, define a global var that we can set and
which is used by the functions to compute the return value.  This
reduces any possible issues with round-tripping a value through the
FFI.

Update the tests appropriately to define the alien variable and set it
before calling the test function.