Rename unsigned_to_bool to int_to_bool
Since the global `test_arg` is an `int`, rename `unsigned_to_bool` to
`int_to_bool` since we're converting an `int` to a `bool` instead of
an `unsigned`.

This makes the name of the function consistent with what is being
converted.  Doesn't affect the current test because we only test with
small non-negative integers.