Modify %alien-funcall deftransform to change the integer return type
When the return type is an integer and the number of bits in the
integer is less than 32, rewrite the return type to be a 32-bit
integer (signed or unsigned according to the original type).

This fixes the issue with calling a function that returns an
alien:boolean, even if we declare it to return an (alien:boolean 8).
Masking still happens.

But this doesn't work for unsigned char or unsigned short; masking has
been elided.  But examination of what gcc does shows that we always
zero extend the result to eax.  So masking isn't really needed in this
case.  Clang does the same thing.