Clean up set-diff code
Remove unused functions having to do with hashing the first list.
Modify list-to-hashtable to only want one list to be hashed and only
returning one value (the hashtable).  Manually inline set-diff-hash2.

Apply Carl's suggest to use ecase instead of cond.

Update tests because the hash table returns the results in the same
order as the list version.