Make +min-list-length-for-hashtable+ a defparameter again
Rename it using the `*foo*` convention too.  It's super-convenient to
be able to set this for testing, which we do in the existing test
suite for this.

It's also convenient to be able to do this when doing timing tests so
we can compare what happens when the list version is used vs when the
hashtable is used.