Replace latin-1 character Latin_Small_Letter_I_With_Diaeresis
In files/math.lisp, the word "naive" is spelled using the character
\Latin_Small_Letter_I_With_Diaeresis.  However, when compiling locally
with a UTF-8 encoding (which is the default), this is invalid.  The
letter needs to be encoded as 2 octets.  I'm too lazy to figure out
how to get emacs to insert the correct encoded character so I'm
replacing it with a simple "i".  This makes the file pure ASCII, so it
should work fine with a UTF-8 encoding.