Allow hashing the shorter of the two lists
Carl suggested we can hash the first list and then run over the second
list and remove the element from the hashtable.  When we're done, if
the hashtable is empty, then the first list is a subset of the second.

This shows some nice speedup when the second list is much longer than
the first.