Fix type-init problem and set min length to 150 for subsetp
I was wrong about `type-init` having hashtables available.  The issue
was `*min-list-length-for-hashtable*` not being defined during
`type-init`.  Also, some timing tests indicate that we need a list
length of about 150 for the hashtable implementation to be faster than
the list implementation.

So, add a new variable `*min-list-length-for-subsetp-hashtable*` with
a default value of 150.  A `cold-init-form` is added so it's available
during `type-init`.

The variable `*allow-hashtable-for-set-functions*` is removed because
we don't need it anymore.

Also, as a test, I set `*min-list-length-for-subsetp-hashtable*` to
10, and we were still able to build lisp without problems.