Add macro PROCESS-SET to handle processing of set functions
All of the (non-destructive) set functions have basically the same
processing code.  The only difference is how to initialize the result
and the test used to determine if an item is to be added to the
result.

We encapsulate this similarity in  the macro `PROCESS-SET` and use it
in `union`, `intersection`, and `set-difference`.