Fix #259:  Use system::*software-version* on linux
Previously, linux-os.lisp was accessing (and defining)
lisp::*software-version*, but that isn't defined.  We should be
using *software-version* from misc.lisp.

To match, *software-type*, we export *software-version* from the
SYSTEM package.

To do this, we need to unintern lisp::*software-version* and that
requires the bootstrap file boot-2023-08.lisp.

Finally, build.sh needs to be updated to look for bootstrap files in
the 21e directory, not 21d.