Fix #259:  Use correct package for *software-version*
It's not super clear what package `*software-version*` should be in,
but there was some confusion.  It's defined in the LISP package (in
misc.lisp), but in linux-os.lisp, which sets it, all the code there is
in the SYSTEM package.

So move `*software-version*` to the SYSTEM package, and update
linux-os.lisp to use it.

It's also weird if `*software-type*` is different, so move it to the
SYSTEM package too, and don't export it anymore, since it seems like
an internal variable not meant to be exposed.

To make these changes, you MUST build with the bootstrap file
boot-2023-08 from the 2023-08 snapshot.  (Probably works if built from
21e, but untested.)