Address review comments.
In `os_get_user_homedir`, use a loop to grow the buffer for `getpwnam_r`
if needed.

The API is changed so that we return a string (or NULL) for the home
directory, and a second arg is the return code which is 0 if
getpwnam_r succeeded (returning the home directory or NULL if the name
does not exist).

Update `unix-get-user-homedir` to use the new API and return two
values: the pathname (or nil) and the return code.