Fix #266: Support ~user in namestrings
First, modify `system:get-user-homedir-pathname` to return a namestring
instead of a pathname.  The current (and only) use is to use the
namestring, so let's not cons a pathname just to get the namestring.

Second, modify `parse-unix-namestring` to look for a leading "~".  If
it's found replace it with the user's home directory.   This is done
via `replace-tilde-user` to do the replacement and
`user-homedir-namestring` to get the user's home directory.