Fix #256: loop for var nil works
`(loop for var nil ...)` should work the same as `(loop for var
of-type nil ...)`.

In `loop-optional-type`, there's a test for `z`.  This is where the
old-style type-specifier goes.  We were failing here because when we
try to find the type-spec via lookup in the loop universe, we fail.
Instead, explicitly check for `z` being `NIL`.  This causes us to pop
the loop source to skip over `nil`.  This is what happens if we had
used `fixnum` instead of `nil`.